backend/requirements.txt
fastapi>=0.127.0
uvicorn[standard]>=0.40.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.21
motor>=3.7.0
pymongo>=4.15.0
python-dotenv>=1.0.0
certifi>=2025.0.0
paho-mqtt>=1.6.1